requirements.txt
moviepy==1.0.3
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
groq==0.4.2
google-generativeai==0.4.1
sarvam==0.0.1
edge-tts==6.1.11
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
